+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to the "emacs-mcx" extension will be documented in this file
 
 ## [Unreleased]
 
+### Add
+- `emacs-mcx.executeCommandWithPrefixArgument` command, #1150.
+
 ## [0.36.6] - 2021-12-18
 ### Fix
 - Internal package updates.
@@ -28,7 +31,7 @@ All notable changes to the "emacs-mcx" extension will be documented in this file
 - Set empty strings to the command field to disable the keybindings instead of leaving the field undefined, #1081.
 
 ## [0.36.0] - 2021-10-04
-## Add
+### Add
 - `C-x r t` (string-rectangle), #1049.
 
 ## [0.35.2] - 2021-09-26